requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
assemblyai==0.17.0
openai==0.28.1
python-dotenv==1.0.0
pydantic==2.4.2 